Task 3028 Otp Bugs
Note:- For Bugs Ticket no :3031,3026,3027,3029,3028,3030

Issue 1: If the entered otp is wrong then we need to refresh the OTP screen as empty fields, currently, the entered OTP is showing event the validation throws
Issue 2: unable to delete only the 2 digit of the OTP post enter the 3 or more digits of the otp
Issue 3: While entering the wrong OTP for the third attempt, the "Account locked message should show instead of showing "Invalid OTP". Because invalid OTP message is showing for 5 sec, then the account locked message is showing, in the meantime, i had entered the email id and attempted to login again.
Issue 4 : Ui content of Otp is modified
Issue 5 : Based on the aws response we are showing expried otp message after 2 min
Issue 6:Resend OTP restricted  maximum limit should be 10 times.  PFB feature description